refactor(e2e): extract accessToken utils
